+++ CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2018-09-21
+## [1.1.0]
 ### Changed
 - Update Changelog to conform with the standards.
 ### Added
 - (DEV) Add support for multiple inputbox in `utilitymanager`.
  > This addition allows for tools to request multiple inputs from the user
 - Add command: mixer
+### Fixed
+- (DEV) Fix lint issues.
 
 
 ## [1.0.2] - 2018-08-22